Convert ulong -> uint32_t

+++ src/common/sndbuf2.h
@@ -18,7 +18,7 @@ class SoundBuffer2 : public SoundSource {
   void Cleanup();
 
   int Get(Sample* dest, int size);
-  ulong GetRate();
+  uint32_t GetRate();
   int GetChannels();
 
   int Fill(int samples);  // バッファに最大 sample 分データを追加
@@ -47,7 +47,7 @@ inline void SoundBuffer2::FillWhenEmpty(bool f) {
   fillwhenempty = f;
 }
 
-inline ulong SoundBuffer2::GetRate() {
+inline uint32_t SoundBuffer2::GetRate() {
   return source ? source->GetRate() : 0;
 }
