+++ src/common/srcbuf.cpp
@@ -50,7 +50,7 @@ SamplingRateConverter::~SamplingRateConverter() {
   Cleanup();
 }
 
-bool SamplingRateConverter::Init(SoundSourceL* _source, int _buffersize, ulong outrate) {
+bool SamplingRateConverter::Init(SoundSourceL* _source, int _buffersize, uint32_t outrate) {
   CriticalSection::Lock lock(cs);
 
   delete[] buffer;
@@ -137,8 +137,8 @@ int SamplingRateConverter::FillMain(int samples) {
 // ---------------------------------------------------------------------------
 //  フィルタを構築
 //
-void SamplingRateConverter::MakeFilter(ulong out) {
-  ulong in = source->GetRate();
+void SamplingRateConverter::MakeFilter(uint32_t out) {
+  uint32_t in = source->GetRate();
 
   // 変換前、変換後レートの比を求める
   // ソースを ic 倍アップサンプリングして LPF を掛けた後
