+++ src/common/srcbuf.h
@@ -13,11 +13,11 @@ class SamplingRateConverter : public SoundSource {
   SamplingRateConverter();
   ~SamplingRateConverter();
 
-  bool Init(SoundSourceL* source, int bufsize, ulong outrate);  // bufsize はサンプル単位
+  bool Init(SoundSourceL* source, int bufsize, uint32_t outrate);  // bufsize はサンプル単位
   void Cleanup();
 
   int Get(Sample* dest, int size);
-  ulong GetRate();
+  uint32_t GetRate();
   int GetChannels();
   int GetAvail();
 
@@ -33,7 +33,7 @@ class SamplingRateConverter : public SoundSource {
   };
 
   int FillMain(int samples);
-  void MakeFilter(ulong outrate);
+  void MakeFilter(uint32_t outrate);
   int Avail();
 
   SoundSourceL* source;
@@ -63,7 +63,7 @@ inline void SamplingRateConverter::FillWhenEmpty(bool f) {
   fillwhenempty = f;
 }
 
-inline ulong SamplingRateConverter::GetRate() {
+inline uint32_t SamplingRateConverter::GetRate() {
   return source ? outputrate : 0;
 }
